Debugs circular imports and db migrations.
