+++ requirements.txt
@@ -1,5 +1,16 @@
-Flask
-Flask-Login
-Flask-JWT-Extended
-Flask-SQLAlchemy
-psycopg2-binary
+blinker==1.8.2
+click==8.1.7
+Flask==3.0.3
+Flask-Cors==4.0.1
+Flask-JWT-Extended==4.6.0
+Flask-Login==0.6.3
+Flask-SQLAlchemy==3.1.1
+greenlet==3.0.3
+itsdangerous==2.2.0
+Jinja2==3.1.4
+MarkupSafe==2.1.5
+psycopg2-binary==2.9.9
+PyJWT==2.8.0
+SQLAlchemy==2.0.31
+typing_extensions==4.12.2
+Werkzeug==3.0.3
